Fix utilization rounding parity for half-hour jobs across both endpoints
Task #34: The two resource-utilization endpoints rounded a timed job's
duration to the nearest 30 min differently at exact half-boundaries. FS
(/resource-utilization) divides by the numeric literal 30.0, so ROUND uses
Postgres' round-half-up. d365crm (/wb/resource-utilization) rounded a double
precision quotient, so ROUND used round-half-to-even. They disagreed at odd
multiples of 15 min (e.g. a 75-min job: 90 in FS vs 60 in d365crm).

Changes:
- artifacts/api-server/src/lib/utilizationSql.ts: cast the d365crm duration
  quotient to ::numeric before ROUND in WB_UTILIZED_MINUTES_SQL so it uses
  round-half-up, matching FS. Added an explanatory comment.
- artifacts/api-server/src/routes/resourceUtilization.test.ts: extended the
  cross-endpoint parity block with exact half-boundary cases (15/45/75/105/135
  min) which the suite previously avoided; both endpoints now agree on
  round-half-up (0.5->1, 2.5->3, 4.5->5).

Verification: vitest suite passes (22 tests), typecheck clean. No behavioural
change away from half-boundaries.

+++ artifacts/api-server/src/lib/utilizationSql.ts
@@ -48,13 +48,19 @@ export const FS_BOOKING_NOT_CANCELLED_SQL = `
 // The CRM booking has no stored duration; the timed duration is derived from the
 // `starttime` / `endtime` timestamptz columns. The booking-status name lives in
 // the OData "FormattedValue" key inside `raw_json`.
+//
+// The duration quotient is cast to `numeric` before ROUND so it uses Postgres'
+// round-half-up behaviour (ROUND on `double precision` uses round-half-to-even).
+// This keeps it in lockstep with FS_UTILIZED_MINUTES_SQL, whose `/ 30.0` divisor
+// already produces a numeric — without the cast the two endpoints disagree at
+// exact half-boundaries (e.g. a 75-min job: 90 in FS vs 60 in d365crm).
 const CRM_BOOKING_STATUS_FV = `b.raw_json->>'_bookingstatus_value@OData.Community.Display.V1.FormattedValue'`;
 
 export const WB_UTILIZED_MINUTES_SQL = `
   CASE
     WHEN b.bookableresourcebookingid IS NULL THEN 0
     WHEN b.starttime IS NULL OR b.endtime IS NULL THEN ${OPEN_ENDED_BOOKING_MINUTES}
-    ELSE ROUND((EXTRACT(EPOCH FROM (b.endtime - b.starttime)) / 60) / 30) * 30
+    ELSE ROUND(((EXTRACT(EPOCH FROM (b.endtime - b.starttime)) / 60) / 30)::numeric) * 30
   END`;
 
 export const WB_BOOKING_NOT_CANCELLED_SQL = `
